Everything pretty much ready. Just need to write it down
Former-commit-id: 9618679585218087c5821ab54397644df63b21b1

+++ src/src/db/DBConnector.java
@@ -100,7 +100,7 @@ public class DBConnector
 		{
 			st = conn.prepareStatement(UPD_CWL_VARS);
 			
-			if(val instanceof Integer)
+			if(val instanceof Long)
 			{
 				st.setLong(1, (long) val);
 				st.setNull(2, java.sql.Types.CHAR);
